requirements.txt
discord.py==1.7.3
discord-py-slash-command==2.4.1
aiohttp==3.7.4.post0
aiosqlite==0.17.0
feedparser==6.0.8
beautifulsoup4==4.9.3
python-dotenv==0.19.2
lxml>=4.9.1